chore: separate subscription assets
