+++ .gitattributes
@@ -1,1 +1,3 @@
 .yarn/** linguist-vendored
+static/**/*.mp4 filter=lfs diff=lfs merge=lfs -text
+static/**/*.webm filter=lfs diff=lfs merge=lfs -text
