+++ static/assets/subscription/videos/EN.mp4
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c28278cf2d5be1a5875b83e0c81871f2d4f5fa1dc957c132790c2c457651b6a1
+size 123293310
