+++ static/assets/subscription/videos/UA.mp4
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:bf3b9035e48fdb7a7d114ca0e9baeaaef79a5c8def8932f63259f52439a01602
+size 113395869
